requirements.txt
imagecodecs
matplotlib
numpy
opencv-python
openslide-python
PyMySQL
pyparsing
tifffile
